refactor: split streaming helpers
